+++ jpx/src/test/java/io/jenetics/jpx/LinkTest.java
@@ -30,6 +30,8 @@ import java.util.function.Supplier;
 
 import org.testng.annotations.Test;
 
+import io.jenetics.jpx.GPX.Version;
+
 /**
  * @author <a href="mailto:franz.wilhelmstoetter@gmail.com">Franz Wilhelmstötter</a>
  */
@@ -42,7 +44,7 @@ public class LinkTest extends XMLStreamTestBase<Link> {
 	}
 
 	@Override
-	protected Params<Link> params(final Random random) {
+	protected Params<Link> params(final Version version, final Random random) {
 		return new Params<>(
 			() -> nextLink(random),
 			Link.READER,